+++ internal/kvstore/table/pack_test.go
@@ -33,7 +33,7 @@ func bval(i int) []byte {
 }
 
 func TestTable_Pack_EncodeDecode(t *testing.T) {
-	size := uint32(1 << 16)
+	size := uint64(1 << 16)
 	tb := New(size)
 
 	timestamp := time.Now().UnixNano()
